refactor some code into a common file
