+++ application-operator/mcagent/mcagent_common.go
@@ -1,0 +1,32 @@
+// Copyright (c) 2021, Oracle and/or its affiliates.
+// Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
+
+package mcagent
+
+import (
+	"context"
+
+	"github.com/go-logr/logr"
+	clustersv1alpha1 "github.com/verrazzano/verrazzano/application-operator/apis/clusters/v1alpha1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+// Syncer contains context for synchronize operations
+type Syncer struct {
+	AdminClient        client.Client
+	LocalClient        client.Client
+	Log                logr.Logger
+	ManagedClusterName string
+	Context            context.Context
+}
+
+// Check if the placement is for this cluster
+func (s *Syncer) isThisCluster(placement clustersv1alpha1.Placement) bool {
+	// Loop through the cluster list looking for the cluster name
+	for _, cluster := range placement.Clusters {
+		if cluster.Name == s.ManagedClusterName {
+			return true
+		}
+	}
+	return false
+}
